refactor: Move HomePage related files into pages/HomePage#52
