+++ src/App.js
@@ -1,7 +1,7 @@
 import { Route, Routes } from "react-router-dom";
 import MobileLayout from "./components/MobileLayout";
 import Login from "./pages/Login";
-import Home from "./pages/Home";
+import HomePage from "./pages/HomePage";
 import Signup from "./pages/SignupPage";
 import Auth from "./pages/Auth";
 
@@ -10,7 +10,7 @@ function App() {
     <MobileLayout>
       <Routes>
         <Route path="/" element={<Login />} />
-        <Route path="/home" element={<Home />} />
+        <Route path="/home" element={<HomePage />} />
         <Route path="/signup" element={<Signup />} />
         <Route path="/auth" element={<Auth />} />
       </Routes>
