+++ src/Pages/HomePage/AttendanceTable.jsx
@@ -8,7 +8,7 @@ import TableContainer from "@mui/material/TableContainer";
 import TableHead from "@mui/material/TableHead";
 import TableRow from "@mui/material/TableRow";
 import Paper from "@mui/material/Paper";
-import useStore from "../store.js";
+import useStore from "../../store.js";
 import { format } from "date-fns";
 import { ko } from "date-fns/locale";
